app: Explicity cast third argument of gimp_enum_radio_box_add to gint
To fix a Apple Clang warning.

gint is what the function on app/widgets/gimpwidgets-utils.h expects.

+++ app/widgets/gimpfilleditor.c
@@ -190,8 +190,8 @@ gimp_fill_editor_constructed (GObject *object)
                                                "pattern-view-size");
       gimp_enum_radio_box_add (GTK_BOX (box), pattern_box,
                                (editor->use_custom_style ?
-                                GIMP_CUSTOM_STYLE_PATTERN :
-                                GIMP_FILL_STYLE_PATTERN),
+                                (gint) GIMP_CUSTOM_STYLE_PATTERN :
+                                (gint) GIMP_FILL_STYLE_PATTERN),
                                FALSE);
     }
